refactor(context): extract getAdminBuildingConfig to utils/adminBuildingConfig.ts
DashboardContext.tsx grew to 4441 lines (+27% in one week). Extract the
400-line getAdminBuildingConfig switch into a pure helper so the context
shrinks (4441 → 4041 lines, -400) and the per-widget validation rules are
independently testable. The context now closes a thin useCallback over
featurePermissions + selectedBuildings and delegates to the pure helper.

Adds 11 unit tests covering legacy building-ID canonicalization, per-widget
validation (reveal-grid columns, drawing width clamp + customColors
padding, countdown viewMode rejection), and unknown-type fallthrough.

Journal: docs/scheduled-tasks/code-structure.md (HIGH item closed).
